Move register decorators to classmethods and import modules in main.py to allow for commands to be written & registered outside of bot.py
